+++ e2e/fixtures/swap-flow-1/popular-assets.json
@@ -1,0 +1,16 @@
+{
+  "request": {
+    "method": "GET",
+    "url": "https://metadata.p.rainbow.me/v1/popular_list?chainId=1"
+  },
+  "response": {
+    "status": 200,
+    "statusText": "OK",
+    "headers": {
+      "content-type": "application/json"
+    },
+    "body": {
+      "data": []
+    }
+  }
+}
