+++ e2e/walletVariables.ts
@@ -50,6 +50,7 @@ export const SWAP_VARIABLES = {
   DAI_MAINNET_ADDRESS: '0x6b175474e89094c44da98b954eedeac495271d0f',
   DAI_MAINNET_ID: '0x6b175474e89094c44da98b954eedeac495271d0f_1',
   ETH_MAINNET_ID: 'eth_1',
+  ETH_HARDHAT_ID: 'eth_31337',
   ETH_OPTIMISM_ID: '0x0000000000000000000000000000000000000000_10',
   GMX_ARBITRUM_ID: '0xfc5a1a6eb076a2c7ad06ed22c90d7e710e35ad0a_42161',
   POL_POLYGON_ID: '0x0000000000000000000000000000000000001010_137',
